+++ R/compute_posterior_m_nn.R
@@ -17,15 +17,13 @@
 #' @param y_obs_ord A vector of observed outcome values. The vector is ordered
 #' as \code{obs_ord}.
 #' @param kernel_fn The covariance function of the GP.
-#' @param n_neighbor The number of nearest neighbors on one side
-#' (see also \code{expand}).
-#' @param expand Scaling factor to determine the total number of nearest
-#' neighbors. The total is \code{2 * expand * n_neighbor}.
+#' @param n_neighbor The number of nearest neighbors on one side.
 #' @param block_size Number of samples included in a computation block.
 #' Mainly used to balance the speed and memory requirement.
 #' Larger \code{block_size} is faster, but requires more memory.
 #'
 #' @return
+#' TODO: The first column is the selected index and the second column is weight.
 #' A two-column matrix. The first column is the weights assigned to each
 #' nearest neighbor. The second column is the corresponding observed outcome
 #' value. The weight in the last row of this matrix is NA and the observed
@@ -41,7 +39,6 @@ compute_posterior_m_nn <- function(hyperparam,
                                    y_obs_ord,
                                    kernel_fn = function(x) exp(-x ^ 2),
                                    n_neighbor = 10,
-                                   expand = 5,
                                    block_size = 1e4) {
 
 
@@ -53,29 +50,30 @@ compute_posterior_m_nn <- function(hyperparam,
   n <- base::length(GPS_w)
 
   # Compute number of blocks
-  n_block <- base::ceiling(n / block_size)
+  # n_block <- base::ceiling(n / block_size)
 
 
   if(w >= obs_ord[nrow(obs_ord),1]){
-    idx_select <- seq(nrow(obs_ord) - expand * n_neighbor + 1, nrow(obs_ord), 1)
+    idx_select <- seq(nrow(obs_ord) - n_neighbor + 1, nrow(obs_ord), 1)
   }else{
+    # which.max returns the index of first TRUE value.
     idx_anchor <- which.max(obs_ord[,1] >= w)
-    idx_start <- max(1, idx_anchor - n_neighbor * expand)
-    idx_end <- min(nrow(obs_ord), idx_anchor + n_neighbor*expand)
+    idx_start <- max(1, idx_anchor - n_neighbor)
+    idx_end <- min(nrow(obs_ord), idx_anchor + n_neighbor)
     if(idx_end == nrow(obs_ord)){
-      idx_select <- seq(idx_end - n_neighbor * 2 * expand + 1, idx_end, 1)
+      idx_select <- seq(idx_end - n_neighbor * 2 + 1, idx_end, 1)
     }else{
-      idx_select <- seq(idx_start, idx_start + n_neighbor * 2 * expand - 1, 1)
+      idx_select <- seq(idx_start, idx_start + n_neighbor * 2 - 1, 1)
     }
   }
 
-  used_obs <- t(t(obs_ord[idx_select, ]) * (1 / sqrt(c(alpha, beta))))
+  used_obs <- t(t(obs_ord[idx_select, ]) * (1 / sqrt(c(beta, alpha))))
   cov_used_inv <- compute_inverse(g_sigma * kernel_fn(as.matrix(dist(used_obs)))
                                   + diag(nrow(used_obs)))
   used_y <- y_obs_ord[idx_select]
 
-  w_obs <- t(t(cbind(w, GPS_w)) * (1/sqrt(c(alpha, beta))))
-  id_all <- split(1:n, ceiling(seq_along(1:n) / n_block))
+  w_obs <- t(t(cbind(w, GPS_w)) * (1/sqrt(c(beta, alpha))))
+  id_all <- split(1:n, ceiling(seq_along(1:n) / block_size))
 
   all_weights <- sapply(id_all, function(id_ind){
     cov_cross <- g_sigma * kernel_fn(spatstat.geom::crossdist(w_obs[id_ind,1],