+++ R/compute_sd_gp.R
@@ -13,12 +13,14 @@
 #'   - Second element: beta
 #'   - Third element: gamma/sigma
 #' @param sigma  A scaler that represents noise.
-#' @param GPS_m  A data.frame of GPS vectors.
-#'   - Column 1: A vector of estimated GPS evaluated at the observed exposure levels.
-#'   - Column 2: Estimated conditional means of the exposure given covariates
-#'               for all samples (e_gps_pred).
-#'   - Column 3: Estimated conditional standard deviation of the exposure given
-#'               covariates for all samples (e_gps_std).
+#' @param GPS_m An S3 gps object including:
+#'   gps: A data.frame of GPS vectors.
+#'     - Column 1: GPS
+#'     - Column 2: Prediction of exposure for covariate of each data sample
+#'     (e_gps_pred).
+#'     - Column 3: Standard deviation of  e_gps (e_gps_std)
+#'   used_params:
+#'     - dnorm_log: TRUE or FLASE
 #' @param kernel_fn The covariance function of GP.
 #'
 #' @keywords internal
@@ -41,12 +43,13 @@ compute_sd_gp <- function(w,
   n <- nrow(scaled_obs)
 
   # Compute GPS for requested w
-  e_gps_pred <- GPS_m$e_gps_pred
-  e_gps_std <- GPS_m$e_gps_std
-  GPS_w <- stats::dnorm(w, mean = e_gps_pred, sd = e_gps_std, log = TRUE)
+  e_gps_pred <- GPS_m$gps$e_gps_pred
+  e_gps_std <- GPS_m$gps$e_gps_std
+  dnorm_log <- GPS_m$used_params$dnorm_log
+  GPS_w <- stats::dnorm(w, mean = e_gps_pred, sd = e_gps_std, log = dnorm_log)
 
   # Compute helper matrix for the new w and corresponding GPS.
-  scaled_w <- cbind(w / sqrt(1 / alpha), GPS_w / sqrt(1 / beta))
+  scaled_w <- cbind(w / sqrt(1 / beta), GPS_w / sqrt(1 / alpha))
 
   scaled_combined <- rbind(scaled_w, scaled_obs)
   Sigma_all <- (g_sigma*kernel_fn(as.matrix(stats::dist(scaled_combined))) +