+++ R/estimate_cerf_gp.R
@@ -11,12 +11,14 @@
 #'   - Column 2: Exposure or treatment (w)
 #'   - Column 3~m: Confounders (C)
 #' @param w A vector of exposure level to compute CERF.
-#' @param GPS_m A data.frame of GPS vectors.
-#'   - Column 1: GPS
-#'   - Column 2: Prediction of exposure for covariate of each data
-#'   sample (e_gps_pred).
-#'   - Column 3: Standard deviation of
-#'   e_gps (e_gps_std)
+#' @param GPS_m An S3 gps object including:
+#'   gps: A data.frame of GPS vectors.
+#'     - Column 1: GPS
+#'     - Column 2: Prediction of exposure for covariate of each data sample
+#'     (e_gps_pred).
+#'     - Column 3: Standard deviation of  e_gps (e_gps_std)
+#'   used_params:
+#'     - dnorm_log: TRUE or FLASE
 #' @param params A list of parameters that is required to run the process.
 #' These parameters include:
 #'   - alpha: A scaling factor for the GPS value.
@@ -66,22 +68,33 @@
 estimate_cerf_gp <- function(data, w, GPS_m, params, nthread = 1,
                              kernel_fn = function(x) exp(-x ^ 2)){
 
-
   # Log system info
   log_system_info()
+
+  # timing the function
+  st_time_gp <- proc.time()
 
   # function call
   fcall <- match.call()
 
   # Double-check input parameters ----------------------------------------------
+  if (any(is.na(data))){
+    stop("At this time, data with missing values is not supported.")
+  }
+
   if (!is.data.frame(data)) {
     stop(paste0("Data should be a data.frame. ",
                 "Current format: ", class(data)[1]))
   }
 
-  if (!is.data.frame(GPS_m)) {
-    stop(paste0("The GPS_m should be a data.frame. ",
+  if (!inherits(GPS_m, "gps")) {
+    stop(paste0("The GPS_m should be a gps class. ",
                 "Current format: ", class(GPS_m)[1]))
+  }
+
+  if (nrow(data)!=length(GPS_m$gps$w)){
+    stop(paste0("Provided Data and GPS object should have the same size. ",
+                "Current sizes: ", nrow(data), " vs ", length(GPS_m$gps$w)))
   }
 
   check_params <- function(my_param, params) {
@@ -96,12 +109,18 @@ estimate_cerf_gp <- function(data, w, GPS_m, params, nthread = 1,
 
   check_params(c("alpha", "beta", "g_sigma", "tune_app"), params)
 
+  # Note: In the package, alpha is used to scale w and beta is used to scale
+  # GPS following the same convention in the method paper.
+
   # TODO: Check values of parameters, too.
 
   # Expand the grid of parameters (alpha, beta, g_sigma) -----------------------
   tune_params <-  expand.grid(getElement(params, "alpha"),
                               getElement(params, "beta"),
                               getElement(params, "g_sigma"))
+
+  logger::log_trace("Number of provided combination of tune parameters: ",
+                    "{nrow(tune_params)}")
 
   if (nrow(tune_params) == 0) {
     stop(paste("Something went wrong with tuning parameters. ",
@@ -119,9 +138,15 @@ estimate_cerf_gp <- function(data, w, GPS_m, params, nthread = 1,
                " is not supported."))
   }
 
+  logger::log_trace("Number of selected combination of tune parameters: ",
+                    "{nrow(tune_params_subset)}")
+
   # Compute m, "confidence interval", and covariate balance for provided
   # hyperparameters. -----------------------------------------------------------
   if(nthread > 1 && nrow(tune_params_subset) > 1) {
+
+    logger::log_trace("Starting a cluster with {nthread} threads ...")
+
     lfp <- get_options("logger_file_path")
 
     # make a cluster
@@ -147,6 +172,7 @@ estimate_cerf_gp <- function(data, w, GPS_m, params, nthread = 1,
                                    })
 
     parallel::stopCluster(cl)
+    logger::log_trace("Cluster with {nthread} threads was terminated.")
   } else if (nrow(tune_params_subset) > 1) {
     tune_res <- apply(tune_params_subset, 1, function(x) {
       compute_m_sigma(hyperparam = x,
@@ -157,6 +183,17 @@ estimate_cerf_gp <- function(data, w, GPS_m, params, nthread = 1,
                       kernel_fn = kernel_fn)
       })
   }
+
+  logger::log_debug("Number of generated tuning results: {length(tune_res)}")
+
+  # Tuning results include:
+  # cb: covariate balance for each confounder. This is the average of all
+  #     all covariate balance for each requested exposure values.
+  # est: Posterior mean for all requested exposure values (vector of
+  #      NA if tuning).
+  # pst: Posterior standard deviation for all requested exposure values (vector
+  #.     of NA if tuning.)
+
 
   # Select the combination of hyperparameters that provides the lowest
   # covariate balance ----------------------------------------------------------
@@ -173,20 +210,33 @@ estimate_cerf_gp <- function(data, w, GPS_m, params, nthread = 1,
                                    tuning = FALSE,
                                    kernel_fn = kernel_fn)
 
-  # Build gp_cerf S3 object
+
+  # Build gp_cerf S3 object ----------------------------------------------------
   result <- list()
   class(result) <- "cerf_gp"
 
-  result$pst_mean <- gp_cerf_final$est
-  result$pst_sd <- gp_cerf_final$pst
-  result$w <- w
+  # Hyper parameters ------------------------
+  optimal_params <- opt_param
+  names(optimal_params) <- c('alpha', 'beta', 'g_sigma')
+  result$optimal_params <- optimal_params
+
+  # Data ------------------------------------
+  posterior <- list()
+  posterior$mean <- gp_cerf_final$est
+  posterior$sd <- gp_cerf_final$pst
+  posterior$w <- w
+  result$posterior <- posterior
+
+  # Covariate balance -----------------------
   result$cb <- gp_cerf_final$cb
-  result$params <- opt_param
+  result$cb_org <- gp_cerf_final$cb_org
+
+  # Function call ---------------------------
   result$fcall <- fcall
 
-  # Add best match to the gp_cerf object
-
-  # Add other useful info form the processing to the gp_cerf object
+  et_time_gp <- proc.time()
+  logger::log_debug("Wall clock time to run estimate_cerf_gp:",
+                    " {(et_time_gp -   st_time_gp)[[3]]} seconds.")
 
   # return gp_cerf S3 object
   invisible(result)