+++ R/estimate_mean_sd_nn.R
@@ -10,16 +10,16 @@
 #' @param w_obs A vector of observed exposure levels.
 #' @param w A vector of exposure levels at which the CERF is estimated.
 #' @param y_obs A vector of observed outcome values.
-#' @param GPS_m A data.frame of GPS vectors.
-#'   - Column 1: GPS
-#'   - Column 2: Prediction of exposure for covariate of each data sample
-#'   (e_gps_pred).
-#'   - Column 3: Standard deviation of  e_gps (e_gps_std)
+#' @param GPS_m An S3 gps object including:
+#'   gps: A data.frame of GPS vectors.
+#'     - Column 1: GPS
+#'     - Column 2: Prediction of exposure for covariate of each data sample
+#'     (e_gps_pred).
+#'     - Column 3: Standard deviation of  e_gps (e_gps_std)
+#'   used_params:
+#'     - dnorm_log: TRUE or FLASE
 #' @param kernel_fn The covariance function of the GP.
-#' @param n_neighbor The number of nearest neighbors on one side
-#' (see also \code{expand}).
-#' @param expand Scaling factor to determine the total number of nearest
-#' neighbors. The total is \code{2 * expand * n_neighbour}.
+#' @param n_neighbor The number of nearest neighbors on one side.
 #' @param block_size The number of samples included in a computation block.
 #' Mainly used to balance the speed and memory requirement. Larger
 #' \code{block_size} is faster, but requires more memory.
@@ -39,7 +39,6 @@ estimate_mean_sd_nn <- function(hyperparam,
                                 GPS_m,
                                 kernel_fn = function(x) exp(-x ^ 2),
                                 n_neighbor = 50,
-                                expand = 2,
                                 block_size = 2e3,
                                 nthread = 1){
 
@@ -48,15 +47,11 @@ estimate_mean_sd_nn <- function(hyperparam,
   logger::log_info("Working on estimating mean and sd using nngp approach ...")
 
 
-  coord_obs <- cbind(w_obs, GPS_m$GPS)
+  coord_obs <- cbind(w_obs, GPS_m$gps$GPS)
 
-  #Remove missing outputs
-  coord_obs <- coord_obs[!is.na(y_obs), ]
-  y_use <- y_obs[!is.na(y_obs)]
-
-  coord_obs_ord <- coord_obs[order(coord_obs[, 1]), ]
-  y_use_ord <- y_use[order(coord_obs[, 1])]
-
+  if (any(is.na(y_obs))){
+    stop("y_obs has missing value(s).")
+  }
 
   lfp <- get_options("logger_file_path")
 
@@ -70,9 +65,9 @@ estimate_mean_sd_nn <- function(hyperparam,
   # export variables and functions to cluster cores
   parallel::clusterExport(cl=cl,
                           varlist = c("w", "GPS_m", "hyperparam",
-                                      "coord_obs_ord", "y_use_ord",
+                                      "coord_obs", "y_obs",
                                       "sigma2", "kernel_fn",
-                                      "n_neighbor", "expand", "block_size",
+                                      "n_neighbor", "block_size",
                                       "compute_posterior_m_nn",
                                       "compute_posterior_sd_nn",
                                       "compute_inverse", "calc_cross"),
@@ -83,17 +78,17 @@ estimate_mean_sd_nn <- function(hyperparam,
                                  w,
                                  function(wi){
     GPS_w <- dnorm(wi,
-                  mean = GPS_m$e_gps_pred,
-                  sd = GPS_m$e_gps_std, log = TRUE)
+                  mean = GPS_m$gps$e_gps_pred,
+                  sd = GPS_m$gps$e_gps_std,
+                  log = GPS_m$used_params$dnorm_log)
 
     val <- compute_posterior_sd_nn(hyperparam = hyperparam,
                                    w = wi,
                                    GPS_w = GPS_w,
-                                   obs_ord = coord_obs_ord,
+                                   obs_ord = coord_obs,
                                    sigma2 = sigma2,
                                    kernel_fn = kernel_fn,
                                    n_neighbor = n_neighbor,
-                                   expand = expand,
                                    block_size = block_size)
     val
   })