+++ functional_tests/ft_generate_synthetic_data.R
@@ -1,0 +1,5 @@
+
+
+set.seed(21)
+synthetic_data <- GPCERF::generate_synthetic_data(sample_size = 1000,
+                                                  gps_spec = 6)