+++ tests/testthat/test-compute_deriv_nn.R
@@ -15,10 +15,9 @@ test_that("compute_deriv_nn works as expected!", {
                                 y_obs = data$Y,
                                 hyperparam = c(0.1,0.2,1),
                                 n_neighbor = 20,
-                                expand = 1,
                                 block_size = 1000)
 
   expect_equal(length(deriv_val), 1L)
-  expect_equal(deriv_val[1, 1], -0.1793515, tolerance = 0.00001)
+  expect_equal(deriv_val[1, 1], 1.558466, tolerance = 0.00001)
   expect_true(is.matrix(deriv_val))
 })