+++ tests/testthat/test-compute_m_sigma.R
@@ -22,7 +22,7 @@ test_that("compute_m_sigma works as expected!", {
    expect_equal(length(gp_cerf), 201L)
    expect_equal(length(w_all), 201L)
    expect_vector(gp_cerf)
-   expect_equal(gp_cerf[10], -26.88945, tolerance = 0.000001)
+   expect_equal(gp_cerf[10], 2.105425, tolerance = 0.000001)
 
    tune_res_t <- compute_m_sigma(hyperparam = c(0.09, 0.09, 10),
                                  data = data,