+++ tests/testthat/test-compute_posterior_m_nn.R
@@ -12,7 +12,6 @@ test_that("compute_posterior_m_nn works as expected.", {
   # Hyperparameter
   hyperparam <- c(0.1, 0.2, 1)
   n_neighbor <- 10
-  expand <- 1
   block_size <- 10000
 
   # Exposure level
@@ -20,11 +19,11 @@ test_that("compute_posterior_m_nn works as expected.", {
 
   # Estimate GPS for the exposure level
   GPS_w <- dnorm(wi,
-                 mean = GPS_m$e_gps_pred,
-                 sd = GPS_m$e_gps_std, log = TRUE)
+                 mean = GPS_m$gps$e_gps_pred,
+                 sd = GPS_m$gps$e_gps_std, log = GPS_m$used_params$dnorm_log)
 
   # Order data for easy selection
-  coord_obs <- cbind(data$treat, GPS_m$GPS)
+  coord_obs <- cbind(data$treat, GPS_m$gps$GPS)
   y_use <- data$Y
 
   obs_ord <- coord_obs[order(coord_obs[, 1]), ]
@@ -36,10 +35,51 @@ test_that("compute_posterior_m_nn works as expected.", {
                                 obs_ord = obs_ord,
                                 y_obs_ord = y_use_ord,
                                 n_neighbor = n_neighbor,
-                                expand = expand,
                                 block_size = block_size)
 
 
   expect_equal(nrow(val), 21L)
   expect_equal(ncol(val), 2L)
+
+  # testing getting the same results with different block size
+
+  val_1 <- compute_posterior_m_nn(hyperparam = hyperparam,
+                                  w = wi,
+                                  GPS_w = GPS_w,
+                                  obs_ord = obs_ord,
+                                  y_obs_ord = y_use_ord,
+                                  n_neighbor = n_neighbor,
+                                  block_size = 40)
+
+  val_2 <- compute_posterior_m_nn(hyperparam = hyperparam,
+                                  w = wi,
+                                  GPS_w = GPS_w,
+                                  obs_ord = obs_ord,
+                                  y_obs_ord = y_use_ord,
+                                  n_neighbor = n_neighbor,
+                                  block_size = 50)
+
+  val_3 <- compute_posterior_m_nn(hyperparam = hyperparam,
+                                  w = wi,
+                                  GPS_w = GPS_w,
+                                  obs_ord = obs_ord,
+                                  y_obs_ord = y_use_ord,
+                                  n_neighbor = n_neighbor,
+                                  block_size = 65)
+
+  val_4 <- compute_posterior_m_nn(hyperparam = hyperparam,
+                                  w = wi,
+                                  GPS_w = GPS_w,
+                                  obs_ord = obs_ord,
+                                  y_obs_ord = y_use_ord,
+                                  n_neighbor = n_neighbor,
+                                  block_size = 10000)
+
+  expect_equal(val_1[15, 2], val_2[15, 2], tolerance = 0.000001)
+  expect_equal(val_1[15, 2], val_3[15, 2], tolerance = 0.000001)
+  expect_equal(val_1[15, 2], val_4[15, 2], tolerance = 0.000001)
+  expect_equal(val_2[12, 2], val_1[12, 2], tolerance = 0.000001)
+  expect_equal(val_2[12, 2], val_3[12, 2], tolerance = 0.000001)
+  expect_equal(val_2[12, 2], val_4[12, 2], tolerance = 0.000001)
+
 })