+++ tests/testthat/test-compute_weight_gp.R
@@ -18,7 +18,10 @@ test_that("multiplication works", {
   e_gps_pred <- predict(e_gps, as.matrix(data[, -(1:2)]))
   e_gps_std <- sd(data$treat - e_gps_pred)
   GPS <- dnorm(data$treat, mean = e_gps_pred, sd = e_gps_std, log = TRUE)
-  GPS_m <- data.frame(GPS, e_gps_pred, e_gps_std)
+
+  GPS_m <- list()
+  GPS_m$gps <- data.frame(GPS, e_gps_pred, e_gps_std)
+  GPS_m$used_params$dnorm_log <- TRUE
 
   # set hyperparameters
   hyperparam <- c(0.1, 0.4, 1)
@@ -27,7 +30,9 @@ test_that("multiplication works", {
   g_sigma <- hyperparam[3]
 
   # Compute scaled observation data and inverse of covariate matrix.
-  scaled_obs = cbind(obs_exposure * sqrt(1 / alpha), GPS * sqrt(1 / beta))
+  scaled_obs = cbind(obs_exposure * sqrt(1 / beta), GPS * sqrt(1 / alpha))
+  colnames(scaled_obs) <- c('w_sc_obs','gps_sc_obs')
+
   sigma_obs = g_sigma * kernel_fn(as.matrix(dist(scaled_obs))) +
                         diag(nrow(scaled_obs))
   inv_sigma_obs <- compute_inverse(sigma_obs)
@@ -41,5 +46,5 @@ test_that("multiplication works", {
                               kernel_fn = kernel_fn)
 
   expect_equal(length(weight$weight), 200L)
-  expect_equal(weight$weight[28], 1.588098e-03, tolerance = 10e-5)
+  expect_equal(weight$weight[28], 0.0002182767, tolerance = 10e-5)
 })