+++ NEWS.md
@@ -1,7 +1,8 @@
 # GPCERF (developing)
 
 ## Changed
-- 
+- The nearest neighbor approach does not get `expand` as an input parameter (`n_neighbor` * `expand` --> `n_neighbor`).
+- The weighted covariate balance now is computed using the wCorr package.
 
 
 # GPCERF 0.2.0 (2023-01-22)